An active_connection file is now also created for WM5-Devices devices. Connected SynCE-Clients now gets informed about connectied WM5-Devices also.
git-svn-id: svn://svn.code.sf.net/p/synce/code/trunk@2352 599ab11a-0111-0410-8a50-d7a5b0953570

+++ vdccm/src/windowscedevice.cpp
@@ -372,7 +372,7 @@ string WindowsCEDevice::getPassword() const
 /*!
     \fn WindowsCEDevice::getBuildNumber() const
  */
-uint16_t WindowsCEDevice::getBuildNumber() const
+uint32_t WindowsCEDevice::getBuildNumber() const
 {
     return buildNumber;
 }
@@ -381,7 +381,7 @@ uint16_t WindowsCEDevice::getBuildNumber() const
 /*!
     \fn WindowsCEDevice::getOsVersion() const
  */
-uint16_t WindowsCEDevice::getOsVersion() const
+uint32_t WindowsCEDevice::getOsVersion() const
 {
     return osVersion;
 }
@@ -390,7 +390,7 @@ uint16_t WindowsCEDevice::getOsVersion() const
 /*!
     \fn WindowsCEDevice::getProcessorType() const
  */
-uint16_t WindowsCEDevice::getProcessorType() const
+uint32_t WindowsCEDevice::getProcessorType() const
 {
     return processorType;
 }
@@ -417,7 +417,7 @@ uint32_t WindowsCEDevice::getPartnerId2() const
 /*!
     \fn WindowsCEDevice::getPort() const
  */
-uint16_t WindowsCEDevice::getPort() const
+uint32_t WindowsCEDevice::getPort() const
 {
     return getRemotePort();
 }
